Use account is_live instead of test_mode when migrating tokens
We can safely assume that an account that's not live will only have
test customers. So, we don't have to take the live mode optimistic
approach everytime. This avoids the loss of test customer data for
test accounts. The optimistic live mode approach is still used when
the account is_live.